% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataAudit-package.R
\docType{package}
\name{dataAudit-package}
\alias{dataAudit}
\alias{dataAudit-package}
\title{dataAudit: Data Audit and Quality Checker}
\description{
A Shiny application for auditing and validating data quality by comparing mined data against historical records.
}
\author{
\strong{Maintainer}: David Kariuki \email{hello@damurka.com}

}
\keyword{internal}
