% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{generate_heatmap}
\alias{generate_heatmap}
\title{Generate a Heatmap of Discrepancies}
\usage{
generate_heatmap(summary)
}
\arguments{
\item{summary}{A `tibble` containing summarized discrepancy proportions by year.}
}
\value{
A `ggplot2` heatmap displaying discrepancy proportions for each indicator over time.
}
\description{
Creates a heatmap visualizing the proportion of discrepancies for each indicator over time.
}
