% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_discrepancies}
\alias{get_discrepancies}
\title{Extract Discrepancies for a Specific Indicator}
\usage{
get_discrepancies(discrepancies, indicator)
}
\arguments{
\item{discrepancies}{A `tibble` containing flagged discrepancies.}

\item{indicator}{A string specifying the name of the indicator to extract discrepancies for.}
}
\value{
A `tibble` containing only the discrepancies for the specified indicator.
}
\description{
Filters the flagged discrepancies for a specific indicator and retains relevant columns.
}
